refactor: make TestTable more flexible to reuse it on other pages
